+++ client/src/components/Popout/Popout.tsx
@@ -4,28 +4,35 @@ import { ReactNode, useEffect, useRef, useState } from 'react'
 import styled, { css } from 'styled-components'
 import { Portal } from '../Portal/Portal'
 
-export type PopoutPosition = 'top-left' | 'bottom-left' | 'bottom-right'
+export type PopoutPosition =
+	| 'top-left'
+	| 'bottom-left'
+	| 'bottom-right'
+	| 'bottom-center'
 
 interface Props {
 	trigger: HTMLElement | null
 	content?: ReactNode
-	showOnHover?: boolean
-	shown?: boolean
 	disableStyle?: boolean
 	position?: PopoutPosition
 	className?: string
+	sticky?: boolean
+	stickyTimeout?: number
 }
 
 export const usePopout = ({
 	trigger,
-	showOnHover = true,
-	shown = true,
 	position = 'top-left',
 	disableStyle,
 	content,
 	className,
+	sticky,
+	stickyTimeout = 200,
 }: Props) => {
-	const [opened, setOpened] = useState(showOnHover ? false : shown)
+	const [triggerHovered, setTriggerHovered] = useState(false)
+	const [contentHovered, setContentHovered] = useState(false)
+
+	const stickyTimeoutRef = useRef<ReturnType<typeof setTimeout>>()
 
 	const [calculatedPosition, setCalculatedPosition] = useState({
 		left: -1000 as number | undefined,
@@ -72,9 +79,16 @@ export const usePopout = ({
 					caretOffset = contentRect.width - rect.width / 2
 					break
 				}
+
+				case 'bottom-center': {
+					left = rect.left + rect.width / 2 - contentRect.width / 2
+					top = rect.bottom + 5
+					caretOffset = contentRect.width / 2
+					break
+				}
 			}
 
-			maxHeight = viewHeight - top - 15
+			maxHeight = viewHeight - top - 30
 
 			setCalculatedPosition({ left, top, maxHeight, caretOffset })
 		}
@@ -82,26 +96,47 @@ export const usePopout = ({
 
 	useEffect(() => {
 		recalculate()
-	}, [opened, shown])
+	}, [triggerHovered])
 
 	useElementEvent(trigger, 'mouseover', () => {
-		setOpened(true)
+		console.log('triggerHovered', triggerHovered)
+
+		if (stickyTimeoutRef.current) {
+			clearTimeout(stickyTimeoutRef.current)
+			stickyTimeoutRef.current = undefined
+		}
+
+		setTriggerHovered(true)
 	})
 
 	useElementEvent(trigger, 'mouseleave', () => {
-		setOpened(false)
+		if (!sticky) {
+			setTriggerHovered(false)
+
+			return
+		}
+
+		if (stickyTimeoutRef.current) {
+			clearTimeout(stickyTimeoutRef.current)
+			stickyTimeoutRef.current = undefined
+		}
+
+		stickyTimeoutRef.current = setTimeout(
+			() => setTriggerHovered(false),
+			stickyTimeout,
+		)
 	})
 
 	useWindowEvent('resize', () => recalculate())
 	useWindowEvent('scroll', () => recalculate(), true)
 
-	console.log({ calculatedPosition })
-
 	return (
 		<>
-			{(opened || (shown && !showOnHover)) && content && (
+			{(triggerHovered || (sticky && contentHovered)) && content && (
 				<Portal>
 					<Container
+						onMouseEnter={() => setContentHovered(true)}
+						onMouseLeave={() => setContentHovered(false)}
 						className={className}
 						ref={contentRef}
 						disableStyle={disableStyle}
@@ -111,133 +146,21 @@ export const usePopout = ({
 							maxHeight: calculatedPosition.maxHeight,
 						}}
 					>
-						{(position === 'bottom-left' || position === 'bottom-right') && (
+						{(position === 'bottom-left' ||
+							position === 'bottom-right' ||
+							position === 'bottom-center') && (
 							<BottomCaret style={{ left: calculatedPosition.caretOffset }} />
 						)}
 						{position === 'top-left' && (
 							<TopCaret style={{ left: calculatedPosition.caretOffset }} />
 						)}
-						{content}
+						<Inner>{content}</Inner>
 					</Container>
 				</Portal>
 			)}
 		</>
 	)
 }
-
-/*
-export const Popout = ({
-	showOnHover = true,
-	shown = true,
-	position = 'top',
-	disableStyle,
-	content,
-	children,
-	className,
-	styleTrigger,
-}: Props) => {
-	const [opened, setOpened] = useState(showOnHover ? false : shown)
-
-	const [calculatedPosition, setCalculatedPosition] = useState({
-		left: -1000 as number | undefined,
-		top: -1000 as number | undefined,
-		caretOffset: -15 as number | undefined,
-		maxHeight: undefined as number | undefined,
-	})
-
-	const triggerRef = useRef<HTMLDivElement>(null)
-	const contentRef = useRef<HTMLDivElement>(null)
-
-	const handleMouseEnter = () => {
-		setOpened(true)
-	}
-
-	const handleMouseLeave = () => {
-		setOpened(false)
-	}
-
-	const recalculate = () => {
-		let top = undefined as number | undefined
-		let left = undefined as number | undefined
-		let maxHeight = undefined as number | undefined
-		let caretOffset = undefined as number | undefined
-
-		const viewHeight = Math.max(
-			document.documentElement.clientHeight,
-			window.innerHeight || 0,
-		)
-
-		if (triggerRef.current && contentRef.current) {
-			const rect = triggerRef.current?.getBoundingClientRect()
-			const contentRect = contentRef.current?.getBoundingClientRect()
-
-			switch (position) {
-				case 'top': {
-					left = rect.left
-					top = rect.top - 5 - contentRect.height
-					caretOffset = rect.width / 2 - 3
-					break
-				}
-
-				case 'bottom': {
-					left = rect.left
-					top = rect.bottom + 5
-					caretOffset = rect.width / 2
-					break
-				}
-			}
-
-			maxHeight = viewHeight - top - 15
-
-			setCalculatedPosition({ left, top, maxHeight, caretOffset })
-		}
-	}
-
-	useEffect(() => {
-		recalculate()
-	}, [opened, shown])
-
-	useWindowEvent('resize', () => recalculate())
-	useWindowEvent('scroll', () => recalculate(), true)
-
-	return (
-		<>
-			<Trigger
-				onMouseEnter={showOnHover ? handleMouseEnter : undefined}
-				onMouseLeave={showOnHover ? handleMouseLeave : undefined}
-				ref={triggerRef}
-				style={styleTrigger}
-				className="tooltip-trigger"
-			>
-				{children}
-			</Trigger>
-
-			{(opened || (shown && !showOnHover)) && content && (
-				<Portal>
-					<Container
-						className={className}
-						ref={contentRef}
-						disableStyle={disableStyle}
-						style={{
-							top: calculatedPosition.top,
-							left: calculatedPosition.left,
-							maxHeight: calculatedPosition.maxHeight,
-						}}
-					>
-						{position === Position.Bottom && (
-							<BottomCaret style={{ left: calculatedPosition.caretOffset }} />
-						)}
-						{position === Position.Top && (
-							<TopCaret style={{ left: calculatedPosition.caretOffset }} />
-						)}
-						{content}
-					</Container>
-				</Portal>
-			)}
-		</>
-	)
-}
-	*/
 
 const Caret = styled.div`
 	content: ' ';
@@ -267,7 +190,17 @@ const Container = styled.div<{ disableStyle?: boolean }>`
 		css`
 			background: ${rgba(props.theme.colors.background, 1)};
 			color: #ddd;
-			padding: 10px;
 			border: 2px solid ${({ theme }) => rgba(theme.colors.border, 1)};
+		`}
+`
+
+const Inner = styled.div<{ disableStyle?: boolean }>`
+	position: relative;
+	overflow: auto;
+
+	${(props) =>
+		!props.disableStyle &&
+		css`
+			padding: 10px;
 		`}
 `
